[svn] Change the logging system (requires recompiling all modules, but no source changes). Treat the original LG_DEBUG, LG_INFO and LG_ERROR and CMDLOG_* as categories, and add some more: LG_RAWDATA, LG_NETWORK (future: netsplit/join), LG_WALLOPS (future). The loglevel directive in atheme.conf now accepts some more values, and only DEBUG and TRACE log all commands. Each value selects a certain subset of categories.

+++ include/global.h
@@ -5,7 +5,7 @@
  *
  * Global data
  *
- * $Id: global.h 7179 2006-11-17 19:58:40Z jilles $
+ * $Id: global.h 7233 2006-11-19 19:25:53Z jilles $
  */
 
 #ifndef _GLOBAL_H
@@ -32,7 +32,7 @@ struct me
   char *mta;                    /* path to mta program                */
   char *numeric;		/* server numeric		      */
 
-  uint8_t loglevel;             /* logging level                      */
+  uint32_t loglevel;            /* logging level                      */
   uint32_t maxfd;               /* how many fds do we have?           */
   uint32_t mdlimit;		/* metadata entry limit		      */
   time_t start;                 /* starting time                      */
